further slice visualizer refactor
